Create Project Code
This program uses a sequential (or linear) search function, the C# Array.BinarySearch() method, and a custom binary search function. An array of 1,000 random numbers between 1 and 10,000 is generated and then the program searches the number from the text box and returns the number and position if found, or not found if it is not in the array.